refactor: batch 패키지 도메인, 리포지토리 common 으로 이전
batch 패키지를 분리한다고 했을 때 도메인, 리포지토리도 같이 사라지면 DB 와의 불일치 문제가 생길 것 같아 common 쪽에서 관리하는게 좋아보여 이전합니다.
