+++ src/main/java/org/bob/siungongsi/batch/scheduler/MessageListener.java
@@ -7,19 +7,19 @@ import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 
 import org.bob.siungongsi.batch.config.SqsProperties;
-import org.bob.siungongsi.batch.domain.GongsiSentStatusEntity;
-import org.bob.siungongsi.batch.domain.ProcessingFailedGongsiEntity;
 import org.bob.siungongsi.batch.event.GongsiMessage;
-import org.bob.siungongsi.batch.repository.GongsiSentStatusRepository;
-import org.bob.siungongsi.batch.repository.ProcessingFailedGongsiRepository;
 import org.bob.siungongsi.batch.service.FileService;
 import org.bob.siungongsi.batch.service.GongsiSummarizer;
 import org.bob.siungongsi.batch.service.OpenDartReader;
 import org.bob.siungongsi.batch.service.ZipFileReader;
 import org.bob.siungongsi.common.domain.CompanyEntity;
 import org.bob.siungongsi.common.domain.GongsiEntity;
+import org.bob.siungongsi.common.domain.GongsiSentStatusEntity;
+import org.bob.siungongsi.common.domain.ProcessingFailedGongsiEntity;
 import org.bob.siungongsi.common.repository.CompanyRepository;
 import org.bob.siungongsi.common.repository.GongsiRepository;
+import org.bob.siungongsi.common.repository.GongsiSentStatusRepository;
+import org.bob.siungongsi.common.repository.ProcessingFailedGongsiRepository;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.ApplicationListener;
 import org.springframework.context.annotation.Profile;
