+++ src/main/java/org/bob/siungongsi/common/domain/GongsiSentStatusEntity.java
@@ -1,7 +1,4 @@
-package org.bob.siungongsi.batch.domain;
-
-import org.bob.siungongsi.common.domain.GongsiEntity;
-import org.bob.siungongsi.common.domain.ModifiableEntity;
+package org.bob.siungongsi.common.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
