+++ src/main/java/org/bob/siungongsi/common/repository/ProcessingFailedGongsiRepository.java
@@ -1,8 +1,8 @@
-package org.bob.siungongsi.batch.repository;
+package org.bob.siungongsi.common.repository;
 
 import java.util.Optional;
 
-import org.bob.siungongsi.batch.domain.ProcessingFailedGongsiEntity;
+import org.bob.siungongsi.common.domain.ProcessingFailedGongsiEntity;
 import org.springframework.data.repository.CrudRepository;
 
 public interface ProcessingFailedGongsiRepository
